+++ src/main/kotlin/com/foodtraceai/model/supplier/SupShipCte.kt
@@ -55,7 +55,7 @@ data class SupShipCte(
 
     // (a)(2) The quantity and unit of measure of the food
     // (e.g., 6 cases, 25 reusable plastic containers, 100 tanks, 200 pounds);
-    val quantity: Short,   // from the supplier
+    val quantity: Int,   // from the supplier
     @Enumerated(EnumType.STRING)
     val unitOfMeasure: UnitOfMeasure,   // from the supplier
 
@@ -114,7 +114,7 @@ data class SupShipCteDto(
     val ftlItem: FtlItem,
     val variety: String,
     val tlcId: Long,
-    val quantity: Short,
+    val quantity: Int,
     val unitOfMeasure: UnitOfMeasure,
     val foodDesc: String,
     val shipToLocationId: Long,
